+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - 'Status' RPC support for all services
 
+## Fixed
+
+- Missing directory service type in API
+
 ## v2.8.3 - 2023-08-22
 
 ## Fixed