+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## v2.9.0 - 2023-11-14
+
 ### Added
 
 - 'Status' RPC support for all services